feat: 환율기반 나라 추천 기능

+++ src/main/java/com/travel/rate/service/ExchgService.java
@@ -18,7 +18,7 @@ public class ExchgService {
         return resExchgDTOS;
     }
 
-    public Map<String, Float> getExchgMap(){
+    public Map<String, Double> getExchgMap(){
         return exchangeUtils.getExchgMap();
     }
